AD-1 refactor: TaskDetailModal 컴포넌트 분리
